requirements.txt
# LLM & AI Framework
langchain>=1.0.0
langchain-openai>=1.0.0
langchain-core>=1.0.0
openai>=1.0.0

# Machine Learning (核心依赖)
xgboost>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0

# Data Processing
pandas>=2.0.0
openpyxl>=3.1.0

# Data Models & Config
pydantic>=2.0.0
python-dotenv>=1.0.0

